pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "capstone"
version = "0.1.0"
description = "Satellite Metadata Search Agent (STAC + Google ADK)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.19.0",
    "python-dotenv>=1.2.1",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "tabulate>=0.9.0",
    "shapely>=2.1.2",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"capstone.aoi" = ["aoi_catalog.json"]
